pyproject.toml
[project]
name = "autoresearch"
version = "0.1.0"
description = "Autonomous pretraining research swarm"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # ML & Training
    "kernels>=0.11.7",
    "torch==2.9.1",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "requests>=2.32.0",
    "rustbpe>=0.1.0",
    "tiktoken>=0.11.0",
    # DonkeyCar & Simulation
    "gym-donkeycar>=22.11.6",
    "gymnasium>=0.29.0",
    "pillow>=10.3.0",
    # Export & Serialization
    "onnx>=1.17.0",
    "onnxruntime>=1.17.0",
    # CLI & Configuration
    "typer[all]>=0.9.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    # Experiment Tracking
    "wandb>=0.16.0",
    # Utilities
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.3.0",
]

[project.scripts]
autoresearch = "autoresearch.cli:main"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["autoresearch"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
autoresearch = ["**/*.yaml", "**/*.yml"]
